Move `ConnectGA4CTATileWidget` to analytics-4.
